Support P4Runtime translation test and bug fix
- Remove P4Runtime translation for the switch_info table since we cannot
translate the CPU port we are using for that table.
- Add new job for Jenkins to test P4Runtime translation.
- Move P4Runtime PTF tests to a separate module
- Fix some typos
