+++ .jenkins/Jenkinsfile
@@ -207,6 +207,12 @@ void runTmPtfTests() {
             sh "TM_DOD=1 ./ptf/run/tm/run \"${PROFILE}\" TEST=int-dod"
             sh "TM_DOD=1 ./ptf/run/tv/run \"${PROFILE}\" TEST=int-dod"
         }
+
+        if (PROFILE.equals("fabric")) {
+          // Special case to test P4Runtime translation feature.
+          // Will be removed once we enable the P4Runtime translation by default.
+          sh "ENABLE_P4RUNTIME_TRANSLATION=1 ./ptf/run/tm/run fabric TEST=p4rt.TableEntryReadWriteTest"
+        }
     }
 }
