+++ ptf/tests/common/fabric_test.py
@@ -1242,6 +1242,12 @@ class FabricTest(P4RuntimeTest):
             priority,
         )
 
+    def read_forwarding_acl_set_output_port(
+      self, priority=DEFAULT_PRIORITY, **matches
+    ):
+        matches = self.build_acl_matches(**matches)
+        return self.read_table_entry("acl.acl", matches, priority)
+
     def read_forwarding_acl_punt_to_cpu(self, eth_type=None, priority=DEFAULT_PRIORITY):
         eth_type_ = stringify(eth_type, 2)
         eth_type_mask_ = stringify(0xFFFF, 2)
